Refactor resuable type utilities for EmojiPicker
